fix(auth): change employeeId type from Integer to Long in JwtUtil for consistent type handling

+++ Tick-It-User-Service/src/main/java/com/harish/TickIt/UserService/util/JwtUtil.java
@@ -35,10 +35,10 @@ public class JwtUtil
 		return ro;
 	}
 
-	public Integer getEmployeeId(String token)
+	public Long getEmployeeId(String token)
 	{
 		SecretKey secretKeyObj = io.jsonwebtoken.security.Keys.hmacShaKeyFor(secretKey.getBytes());
-		return Jwts.parser().verifyWith(secretKeyObj).build().parseSignedClaims(token).getPayload().get("employeeId", Integer.class);
+		return Jwts.parser().verifyWith(secretKeyObj).build().parseSignedClaims(token).getPayload().get("employeeId", Long.class);
 	}
